% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{pnds_example}
\alias{pnds_example}
\title{Get the path of the PNDS toy example files}
\usage{
pnds_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If \code{NULL}, the PNDS toy example files names will be listed.}
}
\value{
A vector with names of all the available PNDS toy example files or the path for specific requested PNDS toy example file.
}
\description{
This function provides the path of the microdata from year 2023 of the PNDS toy example files, loaded with this package.
}
\note{
For more information, visit the survey official website <\url{https://www.ibge.gov.br/estatisticas/sociais/saude/9160-pesquisa-nacional-de-demografia-e-saude.html?=&t=o-que-e}> and consult the other functions of this package, described below.
}
\examples{
pnds_example()
pnds_example(path="exampledata.txt")
pnds_example(path="input_example.txt")
pnds_example(path="dictionaryexample.xls")
pnds_example(path="deflatorexample.xls")
}
\seealso{
\link[PNDSIBGE]{get_pnds} for downloading, labeling, deflating and creating survey design object for PNDS microdata.\cr \link[PNDSIBGE]{read_pnds} for reading PNDS microdata.\cr \link[PNDSIBGE]{pnds_labeller} for labeling categorical variables from PNDS microdata.\cr \link[PNDSIBGE]{pnds_deflator} for adding deflator variables to PNDS microdata.\cr \link[PNDSIBGE]{pnds_design} for creating PNDS survey design object.
}
